big update

+++ pages/edit_profile_script.php
@@ -1,9 +1,11 @@
 <?php
-
-include "loginCrud/db_config.php";
+if (session_status() === PHP_SESSION_NONE) {
+    session_start(); // Ensure the session is started
+}
+include "../loginCrud/db_config.php";
 // upload.php
 // Directory where images will be stored
-$targetDir = "image/upload/";
+$targetDir = "../image/upload/";
 
 // Check if the form was submitted
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
@@ -14,41 +16,38 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $email = $_POST['email'];
 
 
-    // Set default status to true in case no picture is uploaded
-    $upload_picture['status'] = true;
+    $upload_picture = true;
     $user = $server->getSole($connect, $userInfoTable, $id);
     // Check if the file is uploaded
     if (isset($file) && $file['error'] == UPLOAD_ERR_OK) {
         // Call the upload function if a file is uploaded
         if ($user && !empty($user['avatar'])) {
             $oldFile = $targetDir . $user['avatar'];
-
-            // Check if the file exists and delete it
-            if (file_exists($oldFile)) {
+            
+            $upload_picture = $server->upload_picture($connect, $userInfoTable,  $file, $targetDir,$id);
+            if($upload_picture && file_exists($oldFile)){
                 unlink($oldFile); // Deletes the old file
-            }
+            };
         }
-        $upload_picture = $server->upload_picture($connect, $userInfoTable, $id, $file, $targetDir);
+       
     }
 
     $update = $server->update_info($connect, $userInfoTable, $table, $fname, $lname, $email, $id);
-    if ($update['update_info'] && $update['update_user']) {
+    echo '<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>';
+    
+    if ($upload_picture && $update) {
         $user = $server->getSole($connect, $userInfoTable, $id);
         $fullname = ((isset($user['fname']) && $user['fname'] != "ยังไม่ได้ตั้ง") && (isset($user['lname']) && $user['lname'] != "ยังไม่ได้ตั้ง"))
-            ? $user['fname'] . " " . $user['lname']
-            : "ยังไม่ได้ตั้งชื่อ";
+                ? $user['fname'] . " " . $user['lname']
+                : "ยังไม่ได้ตั้งชื่อ";
         $_SESSION['fullname'] = $fullname;
 
-    }
-    ;
-    echo '<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>';
-    if ($upload_picture['status'] && $update) {
         echo '<script>
                 setTimeout(function() {
                     Swal.fire({
                         position: "center",
                         icon: "success",
-                        title: "แก้ไขข้อมูลสําเร็จ",
+                        title: "แก้ไขสำเร็จ",
                         showConfirmButton: true,
                         // timer: 1500
                     }).then(function() {
